+++ rs/execution_environment/tests/dts.rs
@@ -323,10 +323,7 @@ const INSTALL_CODE_INGRESS_COST: u128 = 1_952_000;
 const INSTALL_CODE_EXECUTION_COST: u128 = 5_986_224;
 const NORMAL_INGRESS_COST: u128 = 1_224_000;
 const MAX_EXECUTION_COST: u128 = 6_000_000;
-
-fn actual_execution_cost() -> Cycles {
-    Cycles::new(5_985_224)
-}
+const ACTUAL_EXECUTION_COST: u128 = 5_985_224;
 
 #[test]
 fn dts_install_code_with_concurrent_ingress_sufficient_cycles() {
@@ -334,6 +331,7 @@ fn dts_install_code_with_concurrent_ingress_sufficient_cycles() {
     let install_code_execution_cost = Cycles::new(INSTALL_CODE_EXECUTION_COST);
     let normal_ingress_cost = Cycles::new(NORMAL_INGRESS_COST);
     let max_execution_cost = Cycles::new(MAX_EXECUTION_COST);
+    let actual_execution_cost = Cycles::new(ACTUAL_EXECUTION_COST);
 
     // The initial balance is sufficient to run `install_code` and to send an
     // ingress message concurrently.
@@ -410,7 +408,7 @@ fn dts_install_code_with_concurrent_ingress_sufficient_cycles() {
             - install_code_execution_cost
             - install_code_ingress_cost
             - normal_ingress_cost
-            - actual_execution_cost()
+            - actual_execution_cost
             - config.dirty_page_overhead_cycles(1))
         .get()
     );
@@ -421,6 +419,7 @@ fn dts_install_code_with_concurrent_ingress_insufficient_cycles() {
     let install_code_ingress_cost = Cycles::new(INSTALL_CODE_INGRESS_COST);
     let normal_ingress_cost = Cycles::new(NORMAL_INGRESS_COST);
     let max_execution_cost = Cycles::new(MAX_EXECUTION_COST);
+    let actual_execution_cost = Cycles::new(ACTUAL_EXECUTION_COST);
 
     // The initial balance is not sufficient for both execution and concurrent ingress message.
     let initial_balance = install_code_ingress_cost + normal_ingress_cost.max(max_execution_cost);
@@ -458,7 +457,7 @@ fn dts_install_code_with_concurrent_ingress_insufficient_cycles() {
         env.cycle_balance(canister_id),
         (initial_balance
             - install_code_ingress_cost
-            - actual_execution_cost()
+            - actual_execution_cost
             - config.dirty_page_overhead_cycles(1))
         .get()
     );
@@ -469,6 +468,7 @@ fn dts_install_code_with_concurrent_ingress_and_freezing_threshold_insufficient_
     let install_code_ingress_cost = Cycles::new(INSTALL_CODE_INGRESS_COST);
     let normal_ingress_cost = Cycles::new(NORMAL_INGRESS_COST);
     let max_execution_cost = Cycles::new(MAX_EXECUTION_COST);
+    let actual_execution_cost = Cycles::new(ACTUAL_EXECUTION_COST);
     let freezing_threshold = Cycles::new(10000000);
 
     // The initial balance is not sufficient for both execution and concurrent ingress message.
@@ -508,7 +508,7 @@ fn dts_install_code_with_concurrent_ingress_and_freezing_threshold_insufficient_
         env.cycle_balance(canister_id),
         (initial_balance
             - install_code_ingress_cost
-            - actual_execution_cost()
+            - actual_execution_cost
             - config.dirty_page_overhead_cycles(1))
         .get()
     );