+++ rs/nns/governance/src/governance/tests/mod.rs
@@ -11,7 +11,7 @@ use ic_nervous_system_common::{assert_is_err, assert_is_ok, E8};
 use ic_nervous_system_proto::pb::v1::GlobalTimeOfDay;
 use ic_nns_common::pb::v1::NeuronId;
 #[cfg(feature = "test")]
-use ic_nns_governance_api::pb::v1::CreateServiceNervousSystem as ApiCreateServiceNervousSystem;
+use ic_nns_governance_api::CreateServiceNervousSystem as ApiCreateServiceNervousSystem;
 use ic_protobuf::registry::dc::v1::DataCenterRecord;
 #[cfg(feature = "test")]
 use ic_sns_init::pb::v1::SnsInitPayload;
@@ -22,6 +22,7 @@ use maplit::{btreemap, hashmap};
 use std::{convert::TryFrom, time::Duration};
 
 mod list_neurons;
+mod list_proposals;
 mod neurons_fund;
 mod node_provider_rewards;
 mod stake_maturity;
@@ -146,7 +147,7 @@ mod settle_neurons_fund_participation_mem_tests {
     };
     use ic_base_types::PrincipalId;
     use ic_nns_common::pb::v1::NeuronId;
-    use ic_nns_governance_api::pb::v1::SettleNeuronsFundParticipationResponse;
+    use ic_nns_governance_api::SettleNeuronsFundParticipationResponse;
 
     fn make_dummy_neuron_portion() -> NeuronsFundNeuronPortion {
         NeuronsFundNeuronPortion {
@@ -866,6 +867,7 @@ mod metrics_tests {
                 no: 0,
                 total: 555,
             }),
+            topic: Some(Topic::Governance as i32),
             ..ProposalData::default()
         };
 
@@ -883,6 +885,7 @@ mod metrics_tests {
                 no: 0,
                 total: 1,
             }),
+            topic: Some(Topic::NeuronManagement as i32),
             ..ProposalData::default()
         };
         let governance = Governance::new(
@@ -925,6 +928,7 @@ mod metrics_tests {
                 action: Some(manage_neuron_action.clone()),
                 ..Proposal::default()
             }),
+            topic: Some(Topic::NeuronManagement as i32),
             ..ProposalData::default()
         };
 
@@ -936,6 +940,7 @@ mod metrics_tests {
                 ..Proposal::default()
             }),
             decided_timestamp_seconds: 1,
+            topic: Some(Topic::NeuronManagement as i32),
             ..ProposalData::default()
         };
 
@@ -946,6 +951,7 @@ mod metrics_tests {
                 action: Some(motion_action.clone()),
                 ..Proposal::default()
             }),
+            topic: Some(Topic::Governance as i32),
             ..ProposalData::default()
         };
 
@@ -1290,20 +1296,9 @@ fn test_validate_execute_nns_function() {
                 nns_function: NnsFunction::IcpXdrConversionRate as i32,
                 payload: vec![],
             },
-            "The payload could not be decoded into a UpdateIcpXdrConversionRatePayload: \
-             Cannot parse header "
+            "Proposal is obsolete because NNS_FUNCTION_ICP_XDR_CONVERSION_RATE is obsolete as \
+            conversion rates are now provided by the exchange rate canister automatically."
                 .to_string(),
-        ),
-        (
-            ExecuteNnsFunction {
-                nns_function: NnsFunction::IcpXdrConversionRate as i32,
-                payload: Encode!(&UpdateIcpXdrConversionRatePayload {
-                    xdr_permyriad_per_icp: 0,
-                    ..Default::default()
-                })
-                .unwrap(),
-            },
-            "The proposed rate 0 is below the minimum allowable rate".to_string(),
         ),
         (
             ExecuteNnsFunction {
@@ -1427,14 +1422,6 @@ fn test_validate_execute_nns_function() {
         ExecuteNnsFunction {
             nns_function: NnsFunction::CreateSubnet as i32,
             payload: vec![1u8; PROPOSAL_EXECUTE_NNS_FUNCTION_PAYLOAD_BYTES_MAX],
-        },
-        ExecuteNnsFunction {
-            nns_function: NnsFunction::IcpXdrConversionRate as i32,
-            payload: Encode!(&UpdateIcpXdrConversionRatePayload {
-                xdr_permyriad_per_icp: 101,
-                ..Default::default()
-            })
-            .unwrap(),
         },
         ExecuteNnsFunction {
             nns_function: NnsFunction::AssignNoid as i32,