+++ rs/nns/governance/tests/fixtures/mod.rs
@@ -37,7 +37,7 @@ use ic_nns_governance::{
     },
     storage::reset_stable_memory,
 };
-use ic_nns_governance_api::pb::v1::{
+use ic_nns_governance_api::{
     manage_neuron_response::{self, MergeMaturityResponse},
     ManageNeuronResponse,
 };