+++ rs/nns/governance/CHANGELOG.md
@@ -11,6 +11,29 @@ here were moved from the adjacent `unreleased_changelog.md` file.
 INSERT NEW RELEASES HERE
 
 
+# 2025-05-10: Proposal 136580
+
+http://dashboard.internetcomputer.org/proposal/136580
+
+## Removed
+
+* The `governance_heap_neuron_count` metric is removed as there are no neurons in the heap anymore.
+
+
+# 2025-05-02: Proposal 136427
+
+http://dashboard.internetcomputer.org/proposal/136427
+
+## Changed
+
+* The Governance canister will fetch rewards from the new Node Rewards Canister instead of from Registry.
+
+## Removed
+
+* All the `_pb` methods are removed as they already always panic, as well as decoding the init arg
+  as protobuf.
+
+
 # 2025-04-25: Proposal 136370
 
 http://dashboard.internetcomputer.org/proposal/136370